getting body from msg on queue
I have commented out all the stuff i am no more using. I will add in another commit an other folder which contains a piece of code which continoulsy reads the body of the messges on the queue and sends them to the db.  !!!JUST REALISED NEEED TO TRANSFERT FILES OER IGNORE THIS COMMIT!!